% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{timesheet_raw}
\alias{timesheet_raw}
\title{Sample data for owcontrolling application.}
\format{
A data frame with 150 rows and 6 variables:
\describe{
  \item{costcenter}{chr}
  \item{scheduleid}{int}
  \item{employee}{chr}
  \item{workpackage}{chr}
  \item{startdate}{Date}
  \item{duration}{num}
}
}
\source{
sampled data
}
\usage{
timesheet_raw
}
\description{
A sample dataset containing all necessary variables for the application to
work.
}
\keyword{datasets}
